chore: remove PermissionService from API code (#16153)
